refactor(org): harden archived-org guards across all auth paths
- Extract isOrgArchived to src/core/org-archived.ts; web/lib re-exports
- Replace inline JSON.parse try/catch in resolve-org-from-path with isOrgArchived
- Fix throw-inside-catch anti-pattern in context-factory browser-session path
- Add archived checks to Mesh JWT and API key auth paths (return unauthenticated)
- Rename isOrgNotFoundError → isArchivedOrgError with exact message equality
- Restore toast.success before window.location.href redirect on org delete
